+++ src/models/Settings.php
@@ -103,25 +103,25 @@ class Settings extends Model
                 'label' => Craft::t('sprout-seo', 'General'),
                 'url' => 'sprout-seo/settings/general',
                 'selected' => 'general',
-                'template' => 'sprout-base-seo/settings/general'
+                'template' => 'sprout-seo/settings/general'
             ],
             'sitemaps' => [
                 'label' => Craft::t('sprout-seo', 'Sitemaps'),
                 'url' => 'sprout-seo/settings/sitemaps',
                 'selected' => 'sitemaps',
-                'template' => 'sprout-base-seo/settings/sitemaps'
+                'template' => 'sprout-seo/settings/sitemaps'
             ],
             'redirects' => [
                 'label' => Craft::t('sprout-seo', 'Redirects'),
                 'url' => 'sprout-seo/settings/redirects',
                 'selected' => 'redirects',
-                'template' => 'sprout-base-seo/settings/redirects'
+                'template' => 'sprout-seo/settings/redirects'
             ],
             'advanced' => [
                 'label' => Craft::t('sprout-seo', 'Advanced'),
                 'url' => 'sprout-seo/settings/advanced',
                 'selected' => 'advanced',
-                'template' => 'sprout-base-seo/settings/advanced',
+                'template' => 'sprout-seo/settings/advanced',
             ]
         ];
     }