+++ src/services/Sitemaps.php
@@ -46,7 +46,7 @@ class Sitemaps extends Component
         $customSections = (new Query())
             ->select('*')
             ->from(['{{%sproutseo_sitemaps}}'])
-            ->where('siteId=:siteId', [':siteId' => $siteId])
+            ->where('[[siteId]]=:siteId', [':siteId' => $siteId])
             ->andWhere('type=:type', [':type' => NoSection::class])
             ->all();
 
@@ -108,7 +108,7 @@ class Sitemaps extends Component
         $results = (new Query())
             ->select('*')
             ->from(['{{%sproutseo_sitemaps}}'])
-            ->where(['type' => $type, 'siteId' => $siteId])
+            ->where(['type' => $type, '[[siteId]]' => $siteId])
             ->all();
 
         $sitemapSections = [];
@@ -180,7 +180,9 @@ class Sitemaps extends Component
             return false;
         }
 
-        $sitemapSectionRecord->id = $sitemapSection->id;
+        if ($sitemapSection->id){
+            $sitemapSectionRecord->id = $sitemapSection->id;
+        }
         $sitemapSectionRecord->siteId = $sitemapSection->siteId;
         $sitemapSectionRecord->urlEnabledSectionId = $sitemapSection->urlEnabledSectionId;
         $sitemapSectionRecord->type = $sitemapSection->type;
@@ -288,9 +290,9 @@ class Sitemaps extends Component
         $key = Craft::$app->getSecurity()->generateRandomString(12);
 
         $result = (new Query())
-            ->select('uniqueKey')
+            ->select('[[uniqueKey]]')
             ->from(['{{%sproutseo_sitemaps}}'])
-            ->where(['uniqueKey' => $key])
+            ->where(['[[uniqueKey]]' => $key])
             ->scalar();
 
         if ($result) {